Moves the Upckeywords object to stand along functions in a util module
